added isInState() method and added reset=false parameter to next()
added the
bool isInState(void (*state)())
witch return true or false if the given pointer is equal to current state pointer.
This allows to know from outside the machine if it is in a given state.

also added "reset" parameter (defaulting to "false") to next() method, to allow considering
(if set to "true") a change for the same current state as a real state change. It allows
to reset _runCount and _enterTime even if we did not really change state, e.g. if the state
re-enter itself or if another task re-start it.